requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
pymupdf==1.25.1
ebooklib==0.18
beautifulsoup4==4.12.3
pytesseract==0.3.13
Pillow==11.0.0
nltk==3.9.1
natsort==8.4.0
ollama==0.4.4
rarfile==4.2
